+++ tradingstrategy/pair.py
@@ -123,6 +123,12 @@ class DEXPair:
     #: Pair is listed on an exchange we do not if it is good or not
     flag_unknown_exchange: bool
 
+    #: Denormalised web page and API look up information
+    exchange_slug: Optional[str] = None
+
+    #: Denormalised web page and API look up information
+    pair_slug: Optional[str] = None
+
     #: Block number of the first Uniswap Swap event
     first_swap_at_block_number: Optional[BlockNumber] = None
 
@@ -221,6 +227,13 @@ class DEXPair:
         else:
             exchange_name = f"Exchange #{self.exchange_id}"
         return f"{self.base_token_symbol} - {self.quote_token_symbol}, pair #{self.pair_id} on {exchange_name}"
+
+    def get_trading_pair_page_url(self) -> str:
+        """Get information page for this trading pair."""
+        chain_slug = self.chain_id.get_slug()
+        assert self.exchange_slug
+        assert self.pair_slug
+        return f"https://tradingstrategy.ai/trading-view/{chain_slug}/{self.exchange_slug}/{self.pair_slug}"
 
     def __json__(self, request):
         """Pyramid JSON renderer compatibility.